It's workinggit add -A! Stack corruption errors? Must fix.
